+++ src/main/java/com/gamix/config/WebMvcConfig.java
@@ -23,11 +23,12 @@ public class WebMvcConfig implements WebMvcConfigurer {
 
     @Override
     public void addCorsMappings(CorsRegistry registry) {
-        registry.addMapping("/**")
-                .allowedOrigins(System.getenv("FRONT_END_BASE_URL"))
-                .allowedMethods("GET", "POST", "OPTIONS")
-                .allowedHeaders("*")
-                .exposedHeaders("Authorization")
-                .allowCredentials(true);
+        registry
+            .addMapping("/**")
+            .allowedOrigins(System.getenv("FRONT_END_BASE_URL"))
+            .allowedMethods("GET", "POST", "OPTIONS")
+            .allowedHeaders("*")
+            .exposedHeaders("Authorization")
+            .allowCredentials(true);
     }
 }
