+++ src/main/java/com/gamix/controller/PasswordUserController.java
@@ -1,8 +1,9 @@
 package com.gamix.controller;
 
 import com.gamix.exceptions.ExceptionBase;
-import com.gamix.records.inputs.passwordUserController.SignInPasswordUserInput;
-import com.gamix.records.inputs.passwordUserController.SignUpPasswordUserInput;
+import com.gamix.records.passwordUserController.SessionReturn;
+import com.gamix.records.passwordUserController.SignInPasswordUserInput;
+import com.gamix.records.passwordUserController.SignUpPasswordUserInput;
 import com.gamix.service.PasswordUserService;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
@@ -12,6 +13,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import java.util.HashMap;
 import java.util.Map;
 
 import static com.gamix.utils.ControllerUtils.throwError;
@@ -30,12 +32,19 @@ public class PasswordUserController {
             SignUpPasswordUserInput signUpPasswordUserInput = new SignUpPasswordUserInput(
                     requestBody.get("username"),
                     requestBody.get("email"),
-                    requestBody.get("password"));
+                    requestBody.get("password")
+            );
+
+            SessionReturn session = passwordUserService
+                    .signUpPasswordUser(signUpPasswordUserInput);
+
+            Map<String, String> sessionResponse = new HashMap<>();
+            sessionResponse.put("username", session.username());
+            sessionResponse.put("icon", session.icon());
 
             return ResponseEntity.ok()
-                    .header("Authorization", "Bearer " + passwordUserService
-                            .signUpPasswordUser(signUpPasswordUserInput))
-                    .build();
+                .header("Authorization", "Bearer " + session.token())
+                .body(sessionResponse);
         } catch (ExceptionBase ex) {
             return throwError(ex);
         }
@@ -47,15 +56,20 @@ public class PasswordUserController {
             HttpServletRequest req) throws ExceptionBase {
         try {
             SignInPasswordUserInput signInPasswordUserInput = new SignInPasswordUserInput(
-                    requestBody.get("username"),
-                    requestBody.get("email"),
-                    requestBody.get("password"),
-                    Boolean.parseBoolean(requestBody.get("rememberMe")));
+                requestBody.get("username"),
+                requestBody.get("email"),
+                requestBody.get("password"),
+                Boolean.parseBoolean(requestBody.get("rememberMe"))
+            );
+
+            SessionReturn session = passwordUserService.signInPasswordUser(signInPasswordUserInput);
+            Map<String, String> sessionResponse = new HashMap<>();
+            sessionResponse.put("username", session.username());
+            sessionResponse.put("icon", session.icon());
 
             return ResponseEntity.ok()
-                    .header("Authorization", "Bearer " + passwordUserService
-                            .signInPasswordUser(signInPasswordUserInput))
-                    .build();
+                .header("Authorization", "Bearer " + session.token())
+                .body(sessionResponse);
         } catch (ExceptionBase ex) {
             return throwError(ex);
         }
