+++ src/main/java/com/gamix/records/passwordUserController/SessionReturn.java
@@ -1,0 +1,3 @@
+package com.gamix.records.passwordUserController;
+
+public record SessionReturn(String username, String icon, String token) {}
