+++ src/main/java/com/gamix/records/passwordUserController/SignInPasswordUserInput.java
@@ -1,3 +1,3 @@
-package com.gamix.records.inputs.passwordUserController;
+package com.gamix.records.passwordUserController;
 
 public record SignInPasswordUserInput(String username, String email, String password, boolean rememberMe) {}
