+++ src/main/java/com/gamix/records/passwordUserController/SignUpPasswordUserInput.java
@@ -1,3 +1,3 @@
-package com.gamix.records.inputs.passwordUserController;
+package com.gamix.records.passwordUserController;
 
 public record SignUpPasswordUserInput(String username, String email, String password) {}
