+++ src/main/java/com/gamix/service/UserService.java
@@ -6,7 +6,7 @@ import com.gamix.exceptions.user.UserNotFoundByEmail;
 import com.gamix.exceptions.user.UserNotFoundById;
 import com.gamix.exceptions.user.UserNotFoundByUsername;
 import com.gamix.models.User;
-import com.gamix.records.inputs.userController.PartialUserInput;
+import com.gamix.records.userController.PartialUserInput;
 import com.gamix.repositories.UserRepository;
 import com.gamix.security.JwtManager;
 import jakarta.transaction.Transactional;
